separate PathTracer from PathTracerPanel
