+++ src/main/java/com/marginallyclever/ro3/apps/pathtracer/ProgressListener.java
@@ -1,0 +1,7 @@
+package com.marginallyclever.ro3.apps.pathtracer;
+
+import java.util.EventListener;
+
+public interface ProgressListener extends EventListener {
+    void onProgressUpdate(int progress);
+}
